rvf/backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.7
drf-yasg==1.21.7
django-cors-headers==4.2.0
python-decouple==3.8
pylint==2.17.5
ee==0.2
earthengine-api==0.1.328
google-cloud-storage==2.11.0
pandas==2.0.3
APScheduler==3.10.4
boto3==1.34.69
pdfkit==1.0.0
Pillow==9.3.0
scikit-learn==1.2.2
urllib3==2.2.1
GDAL==3.1.4
geopandas==0.13.2
rasterio==1.3.10
rasterstats==0.19.0
xgboost==2.0.3
